moved export h5m test
